+++ packages/features/ee/workflows/lib/types.ts
@@ -1,6 +1,7 @@
 import type { FORM_SUBMITTED_WEBHOOK_RESPONSES } from "@calcom/app-store/routing-forms/lib/formSubmissionUtils";
+import type { WorkflowPermissions } from "@calcom/features/workflows/repositories/WorkflowPermissionsRepository";
 import type { TimeFormat } from "@calcom/lib/timeFormat";
-import type { Prisma } from "@calcom/prisma/client";
+import type { Membership, Prisma, Workflow as PrismaWorkflow } from "@calcom/prisma/client";
 import type { TimeUnit, WorkflowTemplates, WorkflowTriggerEvents } from "@calcom/prisma/enums";
 import { WorkflowActions } from "@calcom/prisma/enums";
 import type { CalEventResponses, RecurringEvent } from "@calcom/types/Calendar";
@@ -93,3 +94,33 @@ export type ScheduleEmailReminderAction = Extract<
   WorkflowActions,
   "EMAIL_HOST" | "EMAIL_ATTENDEE" | "EMAIL_ADDRESS"
 >;
+
+export type WorkflowListType = PrismaWorkflow & {
+  team: {
+    id: number;
+    name: string;
+    members: Membership[];
+    slug: string | null;
+    logo?: string | null;
+  } | null;
+  steps: WorkflowStep[];
+  activeOnTeams?: {
+    team: {
+      id: number;
+      name?: string | null;
+    };
+  }[];
+  activeOn?: {
+    eventType: {
+      id: number;
+      title: string;
+      parentId: number | null;
+      _count: {
+        children: number;
+      };
+    };
+  }[];
+  readOnly?: boolean;
+  permissions?: WorkflowPermissions;
+  isOrg?: boolean;
+};
