+++ packages/features/users/types/invalidAppCredentials.ts
@@ -1,0 +1,4 @@
+export type InvalidAppCredentialBannerProps = {
+  name: string;
+  slug: string;
+};
